restructured gold layer
